strncat

+++ 0x06-pointers_arrays_strings/1-strncat.c
@@ -11,7 +11,7 @@
 
 char *_strncat(char *dest, char *src, int n)
 {
-	long int i, length;
+	int i, length;
 
 	for (i = 0; dest[i] != '\0'; i++)
 	{
